arch-review: extract byte-identical file-comment logic into useFileComments
Finding 5 from the 2026-08-03 architecture review: the file-comment
subsystem is implemented twice in FileViewerModal.vue and
PinnedFilePanel.vue. This first commit moves the shared state and the
byte-identical functions (verified exact line-for-line in both files)
into web/src/composables/useFileComments.ts, following the
useHoverPinPopover injection pattern.

Shared state: selection/draft anchors, compose/edit text, image refs,
lineCommentMap, csvCellComments, DRAFT_COMMENT_ID. Byte-identical
methods: charOffsetFrom, lineAt, clearHighlights, highlightInMarkdown,
commentIdForLine, isHighlightedLine, commentLineLabel, cancelComment,
cancelEditComment, handleDraftImageUpload, removeDraftImage,
removeEditImage, onCsvCellSelect, startEditComment.

The three anchor helpers (toViewportAnchor, anchorFromElement,
anchorFromCellRect) are near-identical (0.92-0.94) and ship here because
the byte-identical onCsvCellSelect/startEditComment call them; they were
unified via the single containerEl option (modalEl vs mainEl). Highlight
re-application stays in each component (divergent guards) and is routed
back through setApplyHighlights so save/cancel re-render per surface.

Both components call the composable and reach shared state through
comments.*; template-bound names are destructured. The dead 'anchor'
field on FileViewerModal's CommentDraft is dropped (never read).
Tests (36 files / 298 tests) and vue-tsc --noEmit pass.
